Add casts, so that -Wconversion doesn't fail.
Add get_expiry_from_now() as a helper for SAS generation.
Remove unused routines.
Use cJSON_Delete() instead of cJSON_free(), etc.
Add brackets - so can easily add debug to if statements.

+++ src/iotconnect_event.c
@@ -166,7 +166,7 @@ char *iotcl_clone_download_url(IotclEventData data, size_t index) {
         return NULL;
     }
     if ((size_t) cJSON_GetArraySize(urls) > index) {
-        cJSON *url = cJSON_GetArrayItem(urls, index);
+        cJSON *url = cJSON_GetArrayItem(urls, (int) index);
         if (is_valid_string(url)) {
             return iotcl_strdup(url->valuestring);
         } else if (cJSON_IsObject(url)) {
